project03

+++ BE/Host.cs
@@ -9,9 +9,9 @@ namespace BE
     [Serializable]
     public class Host
     {
-        public int HostKey { get; set; }
-        public string FhoneNumber { get; set; }
-        public int BankAccountNumber { get; set; }
+        public long HostKey { get; set; }
+        public string FhoneNumber { get; set; }        
+        public int BankAccountNumber { get; set; }   
         public string PrivateName { get; set; }
         public string FamilyName { get; set; }
         public string MailAddress{ get; set; }
